requirements.txt
ipykernel==6.25.2
matplotlib==3.8.0
numpy==1.23.5
pandas==1.5.3 
seaborn==0.13.0 
pathlib==1.0.1
google-auth==2.23.3
google-cloud-bigquery==3.12.0
google-cloud-storage==2.12.0
scikit-learn==1.3.1
scipy==1.11.3
tensorflow==2.12
pyarrow==13.0.0
db-dtypes==1.1.1
prefect==2.13.7
psycopg2-binary==2.9.9
fastapi==0.103.2
uvicorn==0.23.2
mlflow==1.30.0
python-dotenv==1.0.0
